@@ -1,9 +1,9 @@
-#ifndef CONJUGATE_GRADIENT_OPTIMIZATION
-#define CONJUGATE_GRADIENT_OPTIMIZATION
+#ifndef VMC_CONJUGATEGRADIENTOPTIMIZATION_HPP
+#define VMC_CONJUGATEGRADIENTOPTIMIZATION_HPP
 
-#include "vmc/WFOptimization.hpp"
-#include "vmc/EnergyGradientTargetFunction.hpp"
 #include "nfm/ConjGrad.hpp"
+#include "vmc/EnergyGradientTargetFunction.hpp"
+#include "vmc/WFOptimization.hpp"
 
 #include "mci/MCIntegrator.hpp"
 
@@ -12,24 +12,24 @@
 class ConjugateGradientOptimization: public WFOptimization
 {
 private:
-    long _E_Nmc;
-    long _grad_E_Nmc;
+    const int _E_Nmc;
+    const int _grad_E_Nmc;
 
 public:
-    ConjugateGradientOptimization(WaveFunction * wf, Hamiltonian * H, const long &E_Nmc, const long &grad_E_Nmc, MCI * mci): WFOptimization(wf, H, mci){
-        _E_Nmc = E_Nmc;
-        _grad_E_Nmc = grad_E_Nmc;
-    }
-    virtual ~ConjugateGradientOptimization(){}
+    ConjugateGradientOptimization(WaveFunction * wf, Hamiltonian * H, const int &E_Nmc, const int &grad_E_Nmc, mci::MCI * mci):
+        WFOptimization(wf, H, mci),
+        _E_Nmc(E_Nmc), _grad_E_Nmc(grad_E_Nmc) {}
+
+    ~ConjugateGradientOptimization() override= default;
 
     // optimization
-    void optimizeWF(){
+    void optimizeWF() override{
         // create targetfunction
         EnergyGradientTargetFunction * targetf = new EnergyGradientTargetFunction(_wf, _H, _E_Nmc, _grad_E_Nmc, getMCI());
         // declare the Conjugate Gradient object
         ConjGrad * cjgrad = new ConjGrad(targetf);
         // allocate an array that will contain the wave function variational parameters
-        double * wfpar = new double[_wf->getNVP()];
+        double wfpar[_wf->getNVP()];
         // get the variational parameters
         _wf->getVP(wfpar);
         // set the actual variational parameters as starting point for the Conjugate Gradient algorithm
@@ -40,7 +40,6 @@ public:
         cjgrad->getX(wfpar);
         _wf->setVP(wfpar);
         // free memory
-        delete[] wfpar;
         delete cjgrad;
         delete targetf;
     }
