@@ -1,38 +1,36 @@
-#ifndef ENERGY_GRADIENT_MC_OBSERVABLE
-#define ENERGY_GRADIENT_MC_OBSERVABLE
+#ifndef VMC_ENERGYGRADIENTMCOBSERVABLE_HPP
+#define VMC_ENERGYGRADIENTMCOBSERVABLE_HPP
 
-#include "mci/MCIObservableFunctionInterface.hpp"
-#include "vmc/WaveFunction.hpp"
+#include "mci/ObservableFunctionInterface.hpp"
 #include "vmc/Hamiltonian.hpp"
+#include "vmc/WaveFunction.hpp"
 
 #include <iostream>
 
 
 
-class EnergyGradientMCObservable: public MCIObservableFunctionInterface
+class EnergyGradientMCObservable: public mci::ObservableFunctionInterface
 {
 protected:
-    WaveFunction * _wf;
-    Hamiltonian * _H;
+    WaveFunction * const _wf;
+    Hamiltonian * const _H;
 
+    mci::ObservableFunctionInterface * _clone() const final {
+        return new EnergyGradientMCObservable(_wf, _H);
+    }
 public:
     EnergyGradientMCObservable(WaveFunction * wf, Hamiltonian * H):
-        MCIObservableFunctionInterface(H->getTotalNDim(), 2*wf->getNVP()){
-        _wf = wf;
-        _H = H;
-    }
+        mci::ObservableFunctionInterface(H->getTotalNDim(), 2*wf->getNVP()), _wf(wf), _H(H) {}
 
-    virtual ~EnergyGradientMCObservable(){}
+    ~EnergyGradientMCObservable() override= default;
 
 
-
-    // MCIObservableFunctionInterface implementation
-    void observableFunction(const double * in, double * out){
+    // mci::ObservableFunctionInterface implementation
+    void observableFunction(const double * in, double * out) override{
         // obs[0 : wf->getNVP()-1] = Variational Derivative of the Wave Function
         // obs[ wf->getNVP() : 2*wf->getNVP()-1] = Local Energy times the Variational Derivative of the Wave Function
-        double Hloc;
 
-        Hloc = _H->localPBKineticEnergy(in) + _H->localPotentialEnergy(in);
+        const double Hloc = _H->localPBKineticEnergy(in) + _H->localPotentialEnergy(in);
         for (int i=0; i<_wf->getNVP(); ++i){
             out[i] = _wf->getVD1DivByWF(i);
             out[i+_wf->getNVP()] = out[i] * Hloc;
