@@ -1,15 +1,15 @@
-#ifndef MPIVMC
-#define MPIVMC
+#ifndef VMC_MPIVMC_HPP
+#define VMC_MPIVMC_HPP
 
-#include "mci/MPIMCI.hpp"
 #include "mci/MCIntegrator.hpp"
+#include "mci/MPIMCI.hpp"
 
 #include <iostream>
 #include <string>
 
 namespace MPIVMC
 {
-    int MyRank()
+    inline int MyRank()
     {
         #if USE_MPI==1
         return MPIMCI::myrank();
@@ -18,7 +18,7 @@ namespace MPIVMC
         #endif
     }
 
-    int Size()
+    inline int Size()
     {
         #if USE_MPI==1
         return MPIMCI::size();
@@ -27,7 +27,7 @@ namespace MPIVMC
         #endif
     }
 
-    int Init()
+    inline int Init()
     {
         #if USE_MPI==1
         return MPIMCI::init();
@@ -36,32 +36,32 @@ namespace MPIVMC
         #endif
     }
 
-    void SetSeed(MCI * const mci, const std::string &filename, const int &offset = 0)
+    inline void SetSeed(mci::MCI * mci, const std::string &filename, const int &offset = 0)
     {
         #if USE_MPI==1
-        MPIMCI::setSeed(mci, filename, offset);
+        MPIMCI::setSeed(*mci, filename, offset);
         #else
         std::cout << "Warning: In non-MPI mode MPIVMC::SetSeed ignores the seedfile and sets the seed to offset (default 0)." << std::endl;
         mci->setSeed(offset);
         #endif
     }
 
-    void Integrate(MCI * const mci, const long &Nmc, double * average, double * error, const bool findMRT2step=true, const bool initialdecorrelation=true, const bool randomizeWalkers = false)
+    inline void Integrate(mci::MCI * mci, const int &Nmc, double * average, double * error, const bool findMRT2step=true, const bool initialdecorrelation=true, const bool randomizeWalkers = false)
     {
-        if (randomizeWalkers) mci->newRandomX();
+        if (randomizeWalkers) { mci->newRandomX(); }
         #if USE_MPI==1
-        MPIMCI::integrate(mci, Nmc, average, error, findMRT2step, initialdecorrelation);
+        MPIMCI::integrate(*mci, Nmc, average, error, findMRT2step, initialdecorrelation);
         #else
         mci->integrate(Nmc, average, error, findMRT2step, initialdecorrelation);
         #endif
     }
 
-    void Finalize()
+    inline void Finalize()
     {
         #if USE_MPI==1
         MPIMCI::finalize();
         #endif
     }
-};
+}  // namespace MPIVMC
 
 #endif
