@@ -1,5 +1,5 @@
-#ifndef MULTI_COMPONENT_WAVE_FUNCTION
-#define MULTI_COMPONENT_WAVE_FUNCTION
+#ifndef VMC_MULTICOMPONENTWAVEFUNCTION_HPP
+#define VMC_MULTICOMPONENTWAVEFUNCTION_HPP
 
 
 #include "vmc/WaveFunction.hpp"
@@ -11,28 +11,41 @@
 class MultiComponentWaveFunction: public WaveFunction{
 private:
     std::vector<WaveFunction *> _wfs;
+    bool _flag_newToOld = false; // is set to true on newToOld and set back to false after computeAllDerivates()
+
+    mci::SamplingFunctionInterface * _clone() const final {
+        auto newwf = new MultiComponentWaveFunction(_nspacedim, _npart, _flag_vd1, _flag_d1vd1, _flag_d2vd1);
+        for (auto & wf : _wfs) {
+            newwf->addWaveFunction(wf);
+        }
+        return newwf;
+    }
+
+    // we contain ProtoFunctionInterfaces as members, so we need to implement these:
+    void _newToOld(const mci::WalkerState &wlk) override;
+    void _oldToNew() override;
 
 public:
-    MultiComponentWaveFunction(const int &nspacedim, const int &npart, bool flag_vd1=true, bool flag_d1vd1=true, bool flag_d2vd1=true):
-    WaveFunction(nspacedim, npart, 0, 0, flag_vd1, flag_d1vd1, flag_d2vd1){}
-    virtual ~MultiComponentWaveFunction(){
+    MultiComponentWaveFunction(const int &nspacedim, const int &npart, bool flag_vd1=false, bool flag_d1vd1=false, bool flag_d2vd1=false):
+    WaveFunction(nspacedim, npart, 0, 0, flag_vd1, flag_d1vd1, flag_d2vd1) {}
+    ~MultiComponentWaveFunction() override{
         _wfs.clear();
     }
 
 
     void addWaveFunction(WaveFunction * wf);
 
-    void setVP(const double *vp);
+    void setVP(const double *vp) override;
 
-    void getVP(double *vp);
+    void getVP(double *vp) override;
 
-    void samplingFunction(const double * in, double * out);
+    void protoFunction(const double * in, double * out) override;
 
-    double getAcceptance(const double * protoold, const double * protonew);
+    double acceptanceFunction(const double * protoold, const double * protonew) const override;
 
-    void computeAllDerivatives(const double *x);
+    void computeAllDerivatives(const double *x) override;
 
-    double computeWFValue(const double * protovalues);
+    double computeWFValue(const double * protovalues) const override;
 };
 
 
