@@ -1,9 +1,9 @@
-#ifndef TWO_BODY_JASTROW
-#define TWO_BODY_JASTROW
+#ifndef VMC_TWOBODYJASTROW_HPP
+#define VMC_TWOBODYJASTROW_HPP
 
-#include "vmc/WaveFunction.hpp"
-#include "vmc/TwoBodyPseudoPotential.hpp"
 #include "vmc/ParticleArrayHelper.hpp"
+#include "vmc/TwoBodyPseudoPotential.hpp"
+#include "vmc/WaveFunction.hpp"
 
 #include <stdexcept>
 
@@ -21,13 +21,17 @@ u is a 2-body pseudopotential, and must be implemented as TwoBodyPseudoPotential
 class TwoBodyJastrow: public WaveFunction
 {
 private:
-    TwoBodyPseudoPotential * _u2;
+    TwoBodyPseudoPotential * const _u2;
     ParticleArrayHelper * _pah;
 
+    mci::SamplingFunctionInterface * _clone() const final {
+        return new TwoBodyJastrow(_npart, _u2);
+    }
 public:
     TwoBodyJastrow(const int &npart, TwoBodyPseudoPotential * u2):
-    WaveFunction(u2->getNSpaceDim(), npart, 1, u2->getNVP(), u2->hasVD1(), u2->hasD1VD1(), u2->hasD2VD1()){
-        _u2 = u2;
+        WaveFunction(u2->getNSpaceDim(), npart, 1, u2->getNVP(), u2->hasVD1(), u2->hasD1VD1(), u2->hasD2VD1()),
+        _u2(u2)
+    {
         _pah = new ParticleArrayHelper(u2->getNSpaceDim());
         if (hasD1VD1() && !hasVD1()){
             throw std::invalid_argument( "TwoBodyJastrow derivative d1vd1 requires vd1" );
@@ -36,24 +40,22 @@ public:
             throw std::invalid_argument( "TwoBodyJastrow derivative d2vd1 requires vd1 and d1vd1" );
         }
     }
-    virtual ~TwoBodyJastrow(){
+    ~TwoBodyJastrow() override{
         delete _pah;
     }
 
 
-
-    void setVP(const double *vp){_u2->setVP(vp);}
-    void getVP(double *vp){_u2->getVP(vp);}
-
+    void setVP(const double *vp) override{_u2->setVP(vp);}
+    void getVP(double *vp) override{_u2->getVP(vp);}
 
 
-    void samplingFunction(const double * x, double * protov);
+    void protoFunction(const double * x, double * protov) override;
 
-    double getAcceptance(const double * protoold, const double * protonew);
+    double acceptanceFunction(const double * protoold, const double * protonew) const override;
 
-    void computeAllDerivatives(const double *x);
+    void computeAllDerivatives(const double *x) override;
 
-    double computeWFValue(const double * protovalues);
+    double computeWFValue(const double * protovalues) const override;
 };
 
 
