@@ -1,30 +1,27 @@
-#ifndef WF_OPTIMIZATION
-#define WF_OPTIMIZATION
+#ifndef VMC_WFOPTIMIZATION_HPP
+#define VMC_WFOPTIMIZATION_HPP
 
-#include "vmc/WaveFunction.hpp"
-#include "vmc/Hamiltonian.hpp"
 #include "mci/MCIntegrator.hpp"
+#include "vmc/Hamiltonian.hpp"
+#include "vmc/WaveFunction.hpp"
 
 
 class WFOptimization{
 protected:
-    WaveFunction * _wf;
-    Hamiltonian * _H;
-    MCI * _mci;
+    WaveFunction * const _wf;
+    Hamiltonian * const _H;
+    mci::MCI * const _mci;
 
 public:
-    WFOptimization(WaveFunction * wf, Hamiltonian * H, MCI * mci){
-        _wf=wf;
-        _H=H;
-        _mci = mci;
-    }
+    WFOptimization(WaveFunction * wf, Hamiltonian * H, mci::MCI * mci):
+        _wf(wf), _H(H), _mci(mci) {}
 
-    virtual ~WFOptimization(){}
+    virtual ~WFOptimization()= default;
 
     // getters
     WaveFunction * getWF(){return _wf;}
     Hamiltonian * getH(){return _H;}
-    MCI * getMCI(){return _mci;};
+    mci::MCI * getMCI(){return _mci;};
 
     // optimize the wf
     virtual void optimizeWF() = 0;
