@@ -1,8 +1,8 @@
-#ifndef WAVE_FUNCTION
-#define WAVE_FUNCTION
+#ifndef VMC_WAVEFUNCTION_HPP
+#define VMC_WAVEFUNCTION_HPP
 
-#include "mci/MCISamplingFunctionInterface.hpp"
-#include "mci/MCICallBackOnAcceptanceInterface.hpp"
+#include "mci/CallBackOnMoveInterface.hpp"
+#include "mci/SamplingFunctionInterface.hpp"
 
 #include <iostream>
 
@@ -15,23 +15,30 @@ IMPLEMENTATIONS OF THIS INTERFACE MUST INCLUDE:
     - void getVP(double *vp)
             get the variational parameters
 
-    - void samplingFunction(const double * in, double * out)
-            heritage from MCISamplingFunctionInterface, uses Psi^2
+    - void protoFunction(const double * in, double * out)
+            heritage from mci::SamplingFunctionInterface, uses Psi^2
 
-    - double getAcceptance(const double * protoold, const double * protonew)
-            heritage from MCISamplingFunctionInterface
+    - double acceptanceFunction(const double * protoold, const double * protonew)
+            heritage from mci::SamplingFunctionInterface
+
+    - double computeWFValue(const double * protov)
+            compute the true wave function value from given proto values
 
     - void computeAllDerivatives(const double *x)
             use the setters for derivatives values (setD1DivByWF, setD2DivByWF, etc.)
 
 */
 
-
-class WaveFunction: public MCISamplingFunctionInterface, public MCICallBackOnAcceptanceInterface{
+class WaveFunction: public mci::SamplingFunctionInterface
+{
 protected:
+    const int _nspacedim;
+    const int _npart;
     int _nvp;  //number of variational parameters involved
-    int _npart;
-    int _nspacedim;
+
+    const bool _flag_vd1;
+    const bool _flag_d1vd1;
+    const bool _flag_d2vd1;
 
     double * _d1_divbywf;
     double * _d2_divbywf;
@@ -39,13 +46,8 @@ protected:
     double ** _d1vd1_divbywf;
     double ** _d2vd1_divbywf;
 
-    bool _flag_vd1;
-    bool _flag_d1vd1;
-    bool _flag_d2vd1;
-
 
     void _allocateVariationalDerivativesMemory();
-
 
     // --- getters and setters for the derivatives
     // first derivative divided by the wf
@@ -64,12 +66,19 @@ protected:
     void _setD2VD1DivByWF(const int &id2, const int &ivd1, const double &d2vd1_divbywf){_d2vd1_divbywf[id2][ivd1] = d2vd1_divbywf;}
     double ** _getD2VD1DivByWF(){return _d2vd1_divbywf;}
 
+    void _newToOld(const mci::WalkerState &wlk) override {
+        if (wlk.accepted && wlk.needsObs) {
+            this->computeAllDerivatives(wlk.xnew);
+        }
+    }
+
 public:
-    WaveFunction(const int &nspacedim, const int &npart, const int &ncomp, const int &nvp, bool flag_vd1=true, bool flag_d1vd1=true, bool flag_d2vd1=true);
-    virtual ~WaveFunction();
+    WaveFunction(const int &nspacedim, const int &npart, const int &ncomp/*defines number of proto values*/,
+                 const int &nvp, bool flag_vd1=false, bool flag_d1vd1=false, bool flag_d2vd1=false);
+    ~WaveFunction() override;
 
     int getNSpaceDim(){return _nspacedim;}
-    int getTotalNDim(){return MCISamplingFunctionInterface::getNDim();}
+    int getTotalNDim(){return mci::SamplingFunctionInterface::getNDim();}
     int getNPart(){return _npart;}
     int getNVP(){return _nvp;}
 
@@ -91,12 +100,13 @@ public:
     // wavefunction value, you have to provide a method to reconstruct
     // the wactual wavefunction value (not squared) from the sampling
     // function values, in case it is required e.g. by a wrapper.
-    virtual double computeWFValue(const double * protovalues) = 0;    // --- MUST BE IMPLEMENTED
+    // This method is also used to provide MCI's samplingFunction method.
+    virtual double computeWFValue(const double * protovalues) const = 0;    // --- MUST BE IMPLEMENTED
 
-
-    // --- method herited from MCICallBackOnAcceptanceInterface, that will simply call computeAllDerivatives
-    void callBackFunction(const double *x, const bool flag_observation);
-
+    double samplingFunction(const double protovalues[]) const final { // mainly for use by certain MCI trial moves
+        const double wfval = this->computeWFValue(protovalues);
+        return wfval*wfval; // the sampling function is Psi^2
+    }
 
     // --- getters and setters for the derivatives
     // first derivative divided by the wf
