@@ -1,12 +1,12 @@
 #include "vmc/EuclideanMetric.hpp"
 
-#include <math.h>
+#include <cmath>
 
 
 double EuclideanMetric::dist(const double * r1, const double * r2){
-    double dist = pow(r1[0] - r2[0], 2);
-    for (int i=1; i<getNSpaceDim(); ++i){
-        dist += pow(r1[i] - r2[i], 2);
+    double dist = 0.;
+    for (int i=0; i<getNSpaceDim(); ++i){
+        dist += (r1[i] - r2[i])*(r1[i] - r2[i]);
     }
     return sqrt(dist);
 }
@@ -21,10 +21,10 @@ void EuclideanMetric::distD1(const double * r1, const double * r2, double * out)
 
 void EuclideanMetric::distD2(const double * r1, const double * r2, double * out){
     const double r = dist(r1, r2);
-    const double rpow2 = pow(r, 2);
+    const double rpow2 = r*r;
     const double rpow3 = r*rpow2;
     for (int i=0; i<getNSpaceDim(); ++i){
-        const double ripow2 = pow(r1[i]-r2[i], 2);
+        const double ripow2 = (r1[i]-r2[i])*(r1[i]-r2[i]);
         out[i] = ( rpow2 - ripow2 ) / rpow3;
         out[getNSpaceDim()+i] = ( rpow2 - ripow2 ) / rpow3;
     }
