@@ -4,29 +4,32 @@
 
 
 
-void MultiComponentWaveFunction::computeAllDerivatives(const double *x){
-    for (WaveFunction * wf : _wfs){
-        wf->computeAllDerivatives(x);
+void MultiComponentWaveFunction::computeAllDerivatives(const double x[]){
+    // ! contained wf's computeAllDerivatives() are called already in _newToOld() !
+    if (!_flag_newToOld) { // we need to compute contained derivatives
+        for (auto & wf : _wfs) {
+            wf->computeAllDerivatives(x);
+        }
     }
 
     // first derivative
     for (int i=0; i<getTotalNDim(); ++i){
-        _setD1DivByWF(i, 0.);
+        double d1 = 0.;
         for (WaveFunction * wf : _wfs){
-            _setD1DivByWF(i, getD1DivByWF(i) + wf->getD1DivByWF(i));
+            d1 += wf->getD1DivByWF(i);
         }
+        _setD1DivByWF(i, d1);
     }
     // second derivative
     for (int i=0; i<getTotalNDim(); ++i){
-        _setD2DivByWF(i, 0.);
-        for (WaveFunction * wf : _wfs){
-            _setD2DivByWF(i, getD2DivByWF(i) + wf->getD2DivByWF(i));
-        }
-        for (unsigned int iwf=0; iwf<_wfs.size()-1; ++iwf){
+        double d2 = 0.;
+        for (unsigned int iwf=0; iwf<_wfs.size(); ++iwf){
+            d2 += _wfs[iwf]->getD2DivByWF(i);
             for (unsigned int jwf=iwf+1; jwf<_wfs.size(); ++jwf){
-                _setD2DivByWF(i, getD2DivByWF(i) + 2. * _wfs[iwf]->getD1DivByWF(i) * _wfs[jwf]->getD1DivByWF(i));
+                d2 += 2. * _wfs[iwf]->getD1DivByWF(i) * _wfs[jwf]->getD1DivByWF(i);
             }
         }
+        _setD2DivByWF(i, d2);
     }
     // first variational
     if (hasVD1()){
@@ -77,19 +80,10 @@ void MultiComponentWaveFunction::computeAllDerivatives(const double *x){
                  for (unsigned int jwf=0; jwf<_wfs.size(); ++jwf){
                      if (iwf != jwf){
                          for (int ivp=0; ivp<_wfs[jwf]->getNVP(); ++ivp){
-                             _setD2VD1DivByWF(i, ivp+contvp, getD2VD1DivByWF(i, ivp+contvp) + _wfs[iwf]->getD2DivByWF(i) * _wfs[jwf]->getVD1DivByWF(ivp));
-                         }
-                     }
-                     contvp += _wfs[jwf]->getNVP();
-                 }
-             }
-
-             for (unsigned int iwf=0; iwf<_wfs.size(); ++iwf){
-                 contvp = 0;
-                 for (unsigned int jwf=0; jwf<_wfs.size(); ++jwf){
-                     if (iwf != jwf){
-                         for (int ivp=0; ivp<_wfs[jwf]->getNVP(); ++ivp){
-                             _setD2VD1DivByWF(i, ivp+contvp, getD2VD1DivByWF(i, ivp+contvp) + 2. * _wfs[iwf]->getD1DivByWF(i) * _wfs[jwf]->getD1VD1DivByWF(i, ivp));
+                             _setD2VD1DivByWF(i, ivp+contvp, getD2VD1DivByWF(i, ivp+contvp)
+                                              + _wfs[iwf]->getD2DivByWF(i) * _wfs[jwf]->getVD1DivByWF(ivp)
+                                              + 2. * _wfs[iwf]->getD1DivByWF(i) * _wfs[jwf]->getD1VD1DivByWF(i, ivp)
+                                              );
                          }
                      }
                      contvp += _wfs[jwf]->getNVP();
@@ -111,9 +105,11 @@ void MultiComponentWaveFunction::computeAllDerivatives(const double *x){
              }
         }
     }
+
+    _flag_newToOld = false; // reset flag
 }
 
-double MultiComponentWaveFunction::computeWFValue(const double * protovalues)
+double MultiComponentWaveFunction::computeWFValue(const double * protovalues) const
 {
     double out = 1.;
     int contproto = 0;
@@ -124,21 +120,20 @@ double MultiComponentWaveFunction::computeWFValue(const double * protovalues)
     return out;
 }
 
-double MultiComponentWaveFunction::getAcceptance(const double * protoold, const double * protonew){
+double MultiComponentWaveFunction::acceptanceFunction(const double * protoold, const double * protonew) const {
     double acceptance = 1.;
     int contproto = 0;
     for (WaveFunction * wf : _wfs){
-        acceptance *= wf->getAcceptance(protoold+contproto, protonew+contproto);
+        acceptance *= wf->acceptanceFunction(protoold+contproto, protonew+contproto);
         contproto += wf->getNProto();
     }
     return acceptance;
 }
 
-
-void MultiComponentWaveFunction::samplingFunction(const double * in, double * out){
+void MultiComponentWaveFunction::protoFunction(const double * in, double * out){
     int contproto = 0;
     for (WaveFunction * wf : _wfs){
-        wf->samplingFunction(in, out+contproto);
+        wf->protoFunction(in, out+contproto);
         contproto += wf->getNProto();
     }
 }
@@ -182,4 +177,21 @@ void MultiComponentWaveFunction::addWaveFunction(WaveFunction * wf){
     _wfs.push_back(wf);
     setNProto( getNProto() + wf->getNProto() );
     setNVP( getNVP() + wf->getNVP() );
+}
+
+
+void MultiComponentWaveFunction::_newToOld(const mci::WalkerState &wlk) {
+    for (auto & wf : _wfs) {
+        wf->newToOld(wlk); // here the wf's computeAllDerivatives get called (if accepted&&needsObs)
+    }
+    if (wlk.accepted && wlk.needsObs) {
+        _flag_newToOld = true; // avoid derivative recomputation
+        this->computeAllDerivatives(wlk.xnew);
+    }
+}
+
+void MultiComponentWaveFunction::_oldToNew() {
+    for (auto & wf : _wfs) {
+        wf->oldToNew();
+    }
 }
