@@ -1,46 +1,38 @@
 #include "vmc/SymmetrizerWaveFunction.hpp"
 
-#include <stdexcept>
+#include <algorithm>
 #include <cmath>
+#include <numeric>
 
 unsigned long SymmetrizerWaveFunction::_npart_factorial()
 {
     unsigned long fac = 1;
-    const unsigned long npart = static_cast<unsigned long>(_npart);
-    for (unsigned long i=2; i<=npart; ++i) fac *= i;
+    const auto npart = static_cast<unsigned long>(_npart);
+    for (unsigned long i=2; i<=npart; ++i) { fac *= i; }
+
     return fac;
 }
 
 void SymmetrizerWaveFunction::_swapPositions(double * x, const int &i, const int &j)
 {
     // particle swap (of positions)
-    for (int k=0; k<_nspacedim; ++k) {
-        double xh = x[i*_nspacedim+k];
-        x[i*_nspacedim+k] = x[j*_nspacedim+k];
-
-        x[j*_nspacedim+k] = xh;
-    }
+    std::swap_ranges(x+i*_nspacedim, x+(i+1)*_nspacedim, x+j*_nspacedim);
 }
 
 void SymmetrizerWaveFunction::_swapIndices(int * ids, const int &i, const int &j)
 {
     // particle swap (of indices)
-    for (int k=0; k<_nspacedim; ++k) {
-        int idh = ids[i*_nspacedim+k];
-        ids[i*_nspacedim+k] = ids[j*_nspacedim+k];
-
-        ids[j*_nspacedim+k] = idh;
-    }
+    std::swap_ranges(ids+i*_nspacedim, ids+(i+1)*_nspacedim, ids+j*_nspacedim);
 }
 
 void SymmetrizerWaveFunction::_computeStandardDerivatives(const double * x, const double &normf)
 {
     double protov[_wf->getNProto()];
-    _wf->samplingFunction(x, protov);
+    _wf->protoFunction(x, protov);
     const double wfvalue = _wf->computeWFValue(protov);
     const double normf2 = normf * wfvalue;
 
-    _wf->computeAllDerivatives(x);
+    if (!_flag_newToOld) { _wf->computeAllDerivatives(x); }
 
     const int ndim = getTotalNDim();
     for (int ip=0; ip<ndim; ++ip) {
@@ -74,7 +66,7 @@ void SymmetrizerWaveFunction::_computeStandardDerivatives(const double * x, cons
 void SymmetrizerWaveFunction::_addSwapDerivatives(const double * x, const double &normf, const int * ids)
 {
     double protov[_wf->getNProto()];
-    _wf->samplingFunction(x, protov);
+    _wf->protoFunction(x, protov);
     const double wfvalue = _wf->computeWFValue(protov);
     const double normf2 = normf * wfvalue;
 
@@ -109,7 +101,14 @@ void SymmetrizerWaveFunction::_addSwapDerivatives(const double * x, const double
     }
 }
 
-
+void SymmetrizerWaveFunction::_newToOld(const mci::WalkerState &wlk)
+{
+    _wf->newToOld(wlk); // derivatives are computed here
+    if (wlk.accepted && wlk.needsObs) {
+        _flag_newToOld = true;
+        this->computeAllDerivatives(wlk.xnew);
+    }
+}
 
 void SymmetrizerWaveFunction::computeAllDerivatives(const double * x)
 {
@@ -120,17 +119,15 @@ void SymmetrizerWaveFunction::computeAllDerivatives(const double * x)
     bool isOdd = false; // flip for permutation parity
 
     double protov[_wf->getNProto()];
-    samplingFunction(x, protov);
+    protoFunction(x, protov);
     const double normf = 1. / (_npart_factorial()*computeWFValue(protov)); // normalizing factor
     const double normf2 = -normf; // negative factor for odd permutations in antisym case
 
     // initialize
     iter = 0;
-    for (int i=0; i<_npart; ++i) counts[i] = 0;
-    for (int i=0; i<ndim; ++i) {
-        xh[i] = x[i];
-        idh[i] = i;
-    }
+    std::fill(counts, counts+_npart, 0.);
+    std::copy(x, x+ndim, xh);
+    std::iota(idh, idh+ndim, 0); // range 0..ndim-1
 
     // evaluate unswapped wf
     _computeStandardDerivatives(x, normf);
@@ -150,8 +147,12 @@ void SymmetrizerWaveFunction::computeAllDerivatives(const double * x)
             }
 
             // evaluate and add swap wf
-            if (!_flag_antisymmetric || !isOdd) _addSwapDerivatives(xh, normf, idh);
-            else _addSwapDerivatives(xh, normf2, idh);
+            if (!_flag_antisymmetric || !isOdd) {
+                _addSwapDerivatives(xh, normf, idh);
+            }
+            else {
+                _addSwapDerivatives(xh, normf2, idh);
+            }
 
             ++counts[iter];
             iter = 0;
@@ -161,23 +162,26 @@ void SymmetrizerWaveFunction::computeAllDerivatives(const double * x)
             ++iter;
         }
     }
+
+    _flag_newToOld = false; // reset flag
 }
 
-double SymmetrizerWaveFunction::computeWFValue(const double * protovalues)
+double SymmetrizerWaveFunction::computeWFValue(const double * protovalues) const
 {
-    return protovalues[0]; // sign is important here so we calculate the unsquared wf in samplingFunction
+    return protovalues[0]; // sign is important here so we calculate the unsquared wf in protoFunction
 }
 
-double SymmetrizerWaveFunction::getAcceptance(const double * protoold, const double * protonew)
+double SymmetrizerWaveFunction::acceptanceFunction(const double * protoold, const double * protonew) const
 {
-    double po2 = protoold[0]*protoold[0], pn2 = protonew[0]*protonew[0];
-    if (po2 == 0 && pn2 != 0) return 1.;
-    else if (po2 != 0 && pn2 == 0) return 0.;
-    else return pn2 / po2;
+    const double po2 = protoold[0]*protoold[0];
+    const double pn2 = protonew[0]*protonew[0];
+    if (po2 == 0 && pn2 != 0) { return 1.; }
+    if (po2 != 0 && pn2 == 0) { return 0.; }
+    return pn2 / po2;
 }
 
 
-void SymmetrizerWaveFunction::samplingFunction(const double * in, double * out)
+void SymmetrizerWaveFunction::protoFunction(const double * in, double * out)
 {
     const int ndim = getTotalNDim();
     double outh[_wf->getNProto()], inh[ndim]; // helper arrays for input/output
@@ -187,11 +191,11 @@ void SymmetrizerWaveFunction::samplingFunction(const double * in, double * out)
 
     // initialize
     iter = 0;
-    for (int i=0; i<_npart; ++i) counts[i] = 0.;
-    for (int i=0; i<ndim; ++i) inh[i] = in[i];
+    std::fill(counts, counts+_npart, 0.);
+    std::copy(in, in+ndim, inh);
 
     // evaluate unswapped wf
-    _wf->samplingFunction(in, outh);
+    _wf->protoFunction(in, outh);
     out[0] = normf*_wf->computeWFValue(outh);
 
     // add swapped wfs by heaps algorithm
@@ -207,9 +211,13 @@ void SymmetrizerWaveFunction::samplingFunction(const double * in, double * out)
             }
 
             // evaluate and add swap wf
-            _wf->samplingFunction(inh, outh);
-            if (!_flag_antisymmetric || !isOdd) out[0] += normf*_wf->computeWFValue(outh);
-            else out[0] -= normf*_wf->computeWFValue(outh);
+            _wf->protoFunction(inh, outh);
+            if (!_flag_antisymmetric || !isOdd) {
+                out[0] += normf*_wf->computeWFValue(outh);
+            }
+            else {
+                out[0] -= normf*_wf->computeWFValue(outh);
+            }
 
             ++counts[iter];
             iter = 0;
