@@ -1,12 +1,12 @@
-#ifndef TEST_VMC_FUNCTIONS
-#define TEST_VMC_FUNCTIONS
+#ifndef VMC_TESTVMCFUNCTIONS_HPP
+#define VMC_TESTVMCFUNCTIONS_HPP
 
 #include <cmath>
 
-#include "vmc/WaveFunction.hpp"
+#include "vmc/EuclideanMetric.hpp"
 #include "vmc/Hamiltonian.hpp"
 #include "vmc/TwoBodyPseudoPotential.hpp"
-#include "vmc/EuclideanMetric.hpp"
+#include "vmc/WaveFunction.hpp"
 
 
 class QuadrExponential1D1POrbital: public WaveFunction
@@ -14,11 +14,15 @@ class QuadrExponential1D1POrbital: public WaveFunction
 protected:
     double _a, _b;
 
+    mci::SamplingFunctionInterface * _clone() const final {
+        return new QuadrExponential1D1POrbital(_a, _b);
+    }
+
 public:
     QuadrExponential1D1POrbital(const double a, const double b): WaveFunction(1, 1, 1, 2, true, false, false)
         {_a=a; _b=b;}
 
-    void setVP(const double *in)
+    void setVP(const double *in) final
     {
         _a=in[0];
         //if (_a<0.01) _a=0.01;
@@ -27,24 +31,24 @@ public:
         using namespace std;
         //cout << "change a and b! " << _a << "   " << _b << endl;
     }
-    void getVP(double *out)
+    void getVP(double *out) final
     {
         out[0]=_a;
         out[1]=_b;
     }
 
-    void samplingFunction(const double *in, double *out)
+    void protoFunction(const double *in, double *out) final
     {
         *out = -2.*(_b*(in[0]-_a)*(in[0]-_a));
         std::cout << "out " << *out << std::endl;
     }
 
-    double getAcceptance(const double * protoold, const double * protonew)
+    double acceptanceFunction(const double * protoold, const double * protonew) const final
     {
         return exp(protonew[0]-protoold[0]);
     }
 
-    void computeAllDerivatives(const double *in){
+    void computeAllDerivatives(const double *in) final{
         _setD1DivByWF(0, -2.*_b*(in[0]-_a));
         _setD2DivByWF(0, -2.*_b + (-2.*_b*(in[0]-_a))*(-2.*_b*(in[0]-_a)));
         if (hasVD1()){
@@ -53,7 +57,7 @@ public:
         }
     }
 
-    double computeWFValue(const double * protovalues)
+    double computeWFValue(const double * protovalues) const final
     {
         return exp(0.5*protovalues[0]);
     }
@@ -63,34 +67,39 @@ public:
   Trial Wave Function for N particles in 1 dimension, that uses parameters ai and b, with fixed ai and variational b.
   Psi  =  exp( -b * sum((xi-ai)^2) )
 */
-class QuadrExponential1DNPOrbital: public WaveFunction{
+class QuadrExponential1DNPOrbital: public WaveFunction
+{
 protected:
     double * _a;
     double _b;
     double _bi; // inverse b
+
+    mci::SamplingFunctionInterface * _clone() const final {
+        return new QuadrExponential1DNPOrbital(_npart, _a, _b);
+    }
 
 public:
     QuadrExponential1DNPOrbital(const int &npart, const double * a, const double &b):
     WaveFunction(1, npart, 1, 1, true, true, true), _b(b), _bi(1./b)
     {
         _a=new double[npart];
-        for (int i=0; i<npart; ++i) _a[i] = a[i];
+        for (int i=0; i<npart; ++i) { _a[i] = a[i]; }
     }
 
-    ~QuadrExponential1DNPOrbital()
+    ~QuadrExponential1DNPOrbital() final
     {
         delete [] _a;
     }
 
-    void setVP(const double *in){
+    void setVP(const double *in) final{
         _b=in[0];
     }
 
-    void getVP(double *out){
+    void getVP(double *out) final{
         out[0]=_b;
     }
 
-    void samplingFunction(const double *x, double *out)
+    void protoFunction(const double *x, double *out) final
     {
         *out = 0.;
         for (int i=0; i<_npart; ++i) {
@@ -99,12 +108,12 @@ public:
         *out *= -2.*_b;
     }
 
-    double getAcceptance(const double * protoold, const double * protonew)
+    double acceptanceFunction(const double * protoold, const double * protonew) const final
     {
         return exp(protonew[0]-protoold[0]);
     }
 
-    void computeAllDerivatives(const double *x)
+    void computeAllDerivatives(const double *x) final
     {
         if (hasVD1()){
             _setVD1DivByWF(0, 0);
@@ -127,7 +136,7 @@ public:
         }
     }
 
-    double computeWFValue(const double * protovalues)
+    double computeWFValue(const double * protovalues) const final
     {
         return exp(0.5*protovalues[0]);
     }
@@ -139,35 +148,39 @@ class Gaussian1D1POrbital: public WaveFunction
 protected:
     double _b;
 
+    mci::SamplingFunctionInterface * _clone() const final {
+        return new Gaussian1D1POrbital(_b);
+    }
+
 public:
     explicit Gaussian1D1POrbital(const double b):
     WaveFunction(1, 1, 1, 1, false, false, false){
         _b=b;
     }
 
-    void setVP(const double *in)
+    void setVP(const double *in) final
     {
         _b=*in;
         //if (_b<0.01) _b=0.01;
         using namespace std;
         //cout << "change b! " << _b << endl;
     }
-    void getVP(double *out)
+    void getVP(double *out) final
     {
         *out=_b;
     }
 
-    void samplingFunction(const double *in, double *out)
+    void protoFunction(const double *in, double *out) final
     {
         *out=-2.*_b*(*in)*(*in);
     }
 
-    double getAcceptance(const double * protoold, const double * protonew)
+    double acceptanceFunction(const double * protoold, const double * protonew) const final
     {
         return exp(protonew[0]-protoold[0]);
     }
 
-    void computeAllDerivatives(const double *in){
+    void computeAllDerivatives(const double *in) final{
         _setD1DivByWF(0, -2.*_b*(*in));
         _setD2DivByWF(0, -2.*_b+4.*_b*_b*(*in)*(*in));
         if (hasVD1()){
@@ -175,7 +188,7 @@ public:
         }
     }
 
-    double computeWFValue(const double * protovalues)
+    double computeWFValue(const double * protovalues) const final
     {
         return exp(0.5*protovalues[0]);
     }
@@ -186,20 +199,25 @@ class HarmonicOscillator1D1P: public Hamiltonian
 {
 protected:
     double _w;
+
+    mci::ObservableFunctionInterface * _clone() const final {
+        return new HarmonicOscillator1D1P(_w, _wf);
+    }
+
 public:
     HarmonicOscillator1D1P(const double w, WaveFunction * wf): Hamiltonian(1, 1, wf) {_w=w;}
-    double localPotentialEnergy(const double *r)
+    double localPotentialEnergy(const double *r) final
     {
         return (0.5*_w*_w*(*r)*(*r));
     }
 };
 
 
-class He3u2: public TwoBodyPseudoPotential{
+class He3u2: public TwoBodyPseudoPotential
+{
 /*
     u(r) = b/r^5
 */
-
 private:
     double _b;
 
@@ -208,32 +226,31 @@ public:
     TwoBodyPseudoPotential(em, 1, true, true, true){
         _b = -1.;
     }
-    ~He3u2(){}
 
-    void setVP(const double *vp){_b=vp[0];}
-    void getVP(double *vp){vp[0]=_b;}
+    void setVP(const double *vp) final{_b=vp[0];}
+    void getVP(double *vp) final{vp[0]=_b;}
 
-    double ur(const double &dist){
+    double ur(const double &dist) final{
         return _b/pow(dist, 5);
     }
 
-    double urD1(const double &dist){
+    double urD1(const double &dist) final{
         return -5.*_b/pow(dist, 6);
     }
 
-    double urD2(const double &dist){
+    double urD2(const double &dist) final{
         return 30.*_b/pow(dist, 7);
     }
 
-    void urVD1(const double &dist, double * vd1){
+    void urVD1(const double &dist, double * vd1) final{
         vd1[0] = 1./pow(dist, 5);
     }
 
-    void urD1VD1(const double &dist, double * d1vd1){
+    void urD1VD1(const double &dist, double * d1vd1) final{
         d1vd1[0] = -5./pow(dist, 6);
     }
 
-    void urD2VD1(const double &dist, double * d1vd1){
+    void urD2VD1(const double &dist, double * d1vd1) final{
         d1vd1[0] = 30./pow(dist, 7);
     }
 };
@@ -254,38 +271,37 @@ public:
         _a = a;
         _b = b;
     }
-    ~PolynomialU2(){}
 
-    void setVP(const double *vp){
+    void setVP(const double *vp) final{
         _a=vp[0]; _b=vp[1];
     }
-    void getVP(double *vp){
+    void getVP(double *vp) final{
         vp[0]=_a; vp[1]=_b;
     }
 
-    double ur(const double &r){
+    double ur(const double &r) final{
         return _a * pow(r, 2) + _b * pow(r, 3);
     }
 
-    double urD1(const double &r){
+    double urD1(const double &r) final{
         return 2. * _a * r + 3. * _b * pow(r, 2);
     }
 
-    double urD2(const double &r){
+    double urD2(const double &r) final{
         return 2. * _a + 6. * _b * r;
     }
 
-    void urVD1(const double &r, double * vd1){
+    void urVD1(const double &r, double * vd1) final{
         vd1[0] = r*r;
         vd1[1] = r*r*r;
     }
 
-    void urD1VD1(const double &r, double * d1vd1){
+    void urD1VD1(const double &r, double * d1vd1) final{
         d1vd1[0] = 2.*r;
         d1vd1[1] = 3.*r*r;
     }
 
-    void urD2VD1(const double &r, double * d2vd1){
+    void urD2VD1(const double &r, double * d2vd1) final{
         d2vd1[0] = 2.;
         d2vd1[1] = 6.*r;
     }
@@ -304,39 +320,37 @@ public:
     TwoBodyPseudoPotential(em, 1, true, true, true){
         _K = K;
     }
-    ~FlatU2(){}
 
-    void setVP(const double *vp){
+    void setVP(const double *vp) final{
         _K = vp[0];
     }
-    void getVP(double *vp){
+    void getVP(double *vp) final{
         vp[0] = _K;
     }
 
-    double ur(const double &r){
+    double ur(const double & /*r*/) final{
         return _K;
     }
 
-    double urD1(const double &r){
+    double urD1(const double & /*r*/) final{
         return 0.;
     }
 
-    double urD2(const double &r){
+    double urD2(const double & /*r*/) final{
         return 0.;
     }
 
-    void urVD1(const double &r, double * vd1){
+    void urVD1(const double & /*r*/, double * vd1) final{
         vd1[0] = 1.;
     }
 
-    void urD1VD1(const double &r, double * d1vd1){
+    void urD1VD1(const double & /*r*/, double * d1vd1) final{
         d1vd1[0] = 0.;
     }
 
-    void urD2VD1(const double &r, double * d2vd1){
+    void urD2VD1(const double & /*r*/, double * d2vd1) final{
         d2vd1[0] = 0.;
     }
 };
-
 
 #endif
