@@ -1,7 +1,7 @@
 #include "vmc/EuclideanMetric.hpp"
 
-#include <assert.h>
-#include <math.h>
+#include <cassert>
+#include <cmath>
 #include <iostream>
 #include <random>
 
